requirements.txt
aiohttp==3.9.1
PyYAML==6.0.1
